backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
openai==0.28.1
python-dotenv==1.0.0
supabase==1.0.4
requests
PyJWT==2.8.0
bcrypt==4.0.1
